+++ src/utils/string.rs
@@ -3,7 +3,7 @@ use crate::prelude::*;
 
 pub struct StringUtils {}
 impl StringUtils {
-    pub fn get_difference_percentage_str(number: f32, compare_to: f32) -> Rc<str> {
+    pub fn get_difference_percentage_str(number: f64, compare_to: f64) -> Rc<str> {
         let result = MathUtils::get_difference_percentage(number, compare_to);
         format!(
             "{}{}%",
